+++ lib/karafka/configurators/waterdrop.rb
@@ -4,12 +4,12 @@ module Karafka
     class WaterDrop < Base
       # Sets up a WaterDrop settings
       def setup
-        ::WaterDrop.setup do |config|
-          config.send_messages = true
-          config.connection_pool_size = ::Karafka::App.config.concurrency
-          config.connection_pool_timeout = 1
-          config.kafka_hosts = ::Karafka::App.config.kafka_hosts
-          config.raise_on_failure = true
+        ::WaterDrop.setup do |water_config|
+          water_config.send_messages = true
+          water_config.connection_pool_size = config.concurrency
+          water_config.connection_pool_timeout = 1
+          water_config.kafka_hosts = config.kafka_hosts
+          water_config.raise_on_failure = true
         end
       end
     end